changed unsigned int => size_t

+++ src/impex/gif.cxx
@@ -181,12 +181,12 @@ namespace {
                 b += entries[i][2];
             }
             
-            color[0] = r / entries.size();
-            color[1] = g / entries.size();
-            color[2] = b / entries.size();
+            color[0] = (UInt8)(r / entries.size());
+            color[1] = (UInt8)(g / entries.size());
+            color[2] = (UInt8)(b / entries.size());
         }
         
-        int size() const
+        size_t size() const
         {
             return entries.size();
         }
@@ -196,9 +196,9 @@ namespace {
     void find_color_clusters(void_vector<UInt8> & data, 
             std::vector<ColorCluster> & clusters, void_vector<UInt8> & colors)
     {
-        int count = clusters.size();
-        int size = data.size() / 3;
-        int i, current;
+        size_t count = clusters.size();
+        size_t size = data.size() / 3;
+        size_t i, current;
         for(i=0; i<size; ++i)
         {
             clusters[0].add(data.begin()+3*i);
@@ -206,7 +206,7 @@ namespace {
         
         for(current = 1; current < count; ++current)
         {
-            int largest_index = 0;
+            size_t largest_index = 0;
             for(i=1; i<current; ++i)
             {
                 if(clusters[largest_index] < clusters[i])
@@ -235,16 +235,16 @@ namespace {
     void find_color_indices(void_vector<UInt8> & data, 
            std::vector<ColorCluster> & clusters, void_vector<UInt8> & indices)
     {
-        int count = clusters.size();
+        size_t count = clusters.size();
         UInt8 * base = data.begin();
         
-        int i;
+        size_t i;
         for(i=0; i<count; ++i)
         {
-            for(int j=0; j<clusters[i].size(); ++j)
+            for(size_t j=0; j<clusters[i].size(); ++j)
             {
-                int offset = (clusters[i].entries[j] - base) / 3;
-                indices[offset] = i;
+                size_t offset = (clusters[i].entries[j] - base) / 3;
+                indices[offset] = static_cast<UInt8>(i);
             }
         }
     }
